requirements.txt
apache-airflow-providers-amazon==1.2.0